% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_flex.R
\name{bar_flex}
\alias{bar_flex}
\title{bar_flex}
\usage{
bar_flex(bouts_values, bouts_lengths, f, bts)
}
\arguments{
\item{bouts_values}{...}

\item{bouts_lengths}{...}

\item{f}{...}

\item{bts}{...}
}
\value{
barcodes
}
\description{
bar_flex
}
