% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_generation.R
\name{df_generation}
\alias{df_generation}
\title{df_generation}
\usage{
df_generation(bouts_values, bouts_lengths)
}
\arguments{
\item{bouts_values}{...}

\item{bouts_lengths}{...}
}
\value{
df
}
\description{
df_generation
}
